requirements.txt
matplotlib
gmalthgtparser
opencv-python
scipy
numpy-stl
vtkplotlib